+++ tests/utils/test_delete.py
@@ -2,6 +2,7 @@
 
 from cgbeacon2.utils.delete import delete_dataset
 
+
 def test_delete_dataset_none_id():
     """Test trigger error in delete_dataset"""